refactor: migrate switch to header from dropdown
